pyproject.toml
[project]
name = "ctgov-compliance-web"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "amqp==5.3.1",
    "annotated-types==0.7.0",
    "anyio==4.9.0",
    "bidict==0.23.1",
    "billiard==4.2.1",
    "blinker==1.9.0",
    "boto3==1.38.32",
    "botocore==1.38.32",
    "celery==5.5.3",
    "certifi==2025.4.26",
    "charset-normalizer==3.4.2",
    "click==8.2.1",
    "click-didyoumean==0.3.1",
    "click-plugins==1.1.1",
    "click-repl==0.3.0",
    "distro==1.9.0",
    "djlint>=1.36.4",
    "dnspython==2.7.0",
    "ecdsa==0.19.1",
    "email-validator==2.2.0",
    "flask==3.1.1",
    "flask-login==0.6.3",
    "flask-socketio==5.5.1",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf==1.2.2",
    "gevent==25.5.1",
    "gevent-websocket==0.10.1",
    "greenlet==3.2.3",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "idna==3.10",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jmespath==1.0.1",
    "kombu==5.5.4",
    "markupsafe==3.0.2",
    "numpy>=2.3.1",
    "openai==1.85.0",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-exporter-gcp-monitoring>=1.9.0a0",
    "opentelemetry-exporter-gcp-trace>=1.9.0",
    "opentelemetry-instrumentation-flask>=0.57b0",
    "opentelemetry-instrumentation-psycopg2>=0.57b0",
    "opentelemetry-instrumentation-requests>=0.57b0",
    "opentelemetry-propagator-gcp>=1.9.0",
    "opentelemetry-sdk>=1.36.0",
    "packaging==25.0",
    "pandas>=2.3.0",
    "prompt-toolkit==3.0.51",
    "psycopg2-binary",
    "pydantic==2.11.5",
    "pydantic-core==2.33.2",
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.1.0",
    "python-engineio==4.12.2",
    "python-http-client==3.3.7",
    "python-socketio==5.13.0",
    "pytz==2025.2",
    "redis==6.2.0",
    "requests==2.32.4",
    "s3transfer==0.13.0",
    "sendgrid==6.12.3",
    "setuptools==80.9.0",
    "simple-websocket==1.1.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "stripe==12.2.0",
    "tqdm==4.67.1",
    "typing-extensions==4.14.0",
    "typing-inspection==0.4.1",
    "tzdata==2025.2",
    "urllib3==2.4.0",
    "vine==5.1.0",
    "wcwidth==0.2.13",
    "werkzeug==3.1.3",
    "wsproto==1.2.0",
    "wtforms==3.2.1",
    "zope-event==5.0",
    "zope-interface==7.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["web"]

[project.scripts]
devserver = "web.devserver:main"
